+++ packages/application/src/plugin/queries/get-plugin/get-plugin.result.ts
@@ -5,6 +5,7 @@ export interface GetPluginResult {
   readonly name: string;
   readonly version: string;
   readonly icon: string;
+  readonly installPath: string;
   readonly state: PluginRuntimeState;
   readonly enabled: boolean;
 }